demo_clients/demo_client_streamlit_emmanuel/requirements.txt
streamlit
streamlit-option-menu
opencv-python
Pillow
